requirements-dev.txt
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
asyncio-dgram==2.1.2 ; python_version >= "3.10" and python_version < "4"
attrs==22.2.0 ; python_version >= "3.10" and python_version < "4.0"
black==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
colorlog==4.8.0 ; python_version >= "3.10" and python_version < "4.0"
dacite==1.8.0 ; python_version >= "3.10" and python_version < "4.0"
dnspython==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.92.0 ; python_version >= "3.10" and python_version < "4.0"
flake8-bugbear==23.2.13 ; python_version >= "3.10" and python_version < "4.0"
flake8==6.0.0 ; python_version >= "3.10" and python_version < "4.0"
gitdb==4.0.10 ; python_version >= "3.10" and python_version < "4.0"
gitpython==3.1.31 ; python_version >= "3.10" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.10" and python_version < "4.0" and platform_machine == "aarch64" or python_version >= "3.10" and python_version < "4.0" and platform_machine == "ppc64le" or python_version >= "3.10" and python_version < "4.0" and platform_machine == "x86_64" or python_version >= "3.10" and python_version < "4.0" and platform_machine == "amd64" or python_version >= "3.10" and python_version < "4.0" and platform_machine == "AMD64" or python_version >= "3.10" and python_version < "4.0" and platform_machine == "win32" or python_version >= "3.10" and python_version < "4.0" and platform_machine == "WIN32"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
isort==5.12.0 ; python_version >= "3.10" and python_version < "4.0"
mccabe==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
mcstatus==10.0.1 ; python_version >= "3.10" and python_version < "4"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
mypy==0.991 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.10" and python_version < "4.0"
pathspec==0.11.0 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==3.0.0 ; python_version >= "3.10" and python_version < "4.0"
pycodestyle==2.10.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.5 ; python_version >= "3.10" and python_version < "4.0"
pyflakes==3.0.1 ; python_version >= "3.10" and python_version < "4.0"
pysen==0.10.3 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.21.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
schedule==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
smmap==5.0.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==2.0.4 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.25.0 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
tomlkit==0.11.6 ; python_version >= "3.10" and python_version < "4.0"
types-pyyaml==6.0.12.6 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "4.0"
unidiff==0.7.4 ; python_version >= "3.10" and python_version < "4.0"
uvicorn[standard]==0.20.0 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.17.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
watchfiles==0.18.1 ; python_version >= "3.10" and python_version < "4.0"
websockets==10.4 ; python_version >= "3.10" and python_version < "4.0"
